requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
markdown2
python-dotenv
gunicorn